nano-summarizer-website/requirements.txt
google-generativeai
matplotlib
seaborn
